requirements.txt
numpy
matplotlib
pandas
numba
joblib
numpy-stl
tqdm
scipy
seaborn
scikit-learn
pyvista
vtk
pyyaml
h5py
spiceypy